[Game] HLSDK 1.0 static link: g_engfuncs table + save/restore export registry

+++ src/halflife/animation.cpp
@@ -17,9 +17,10 @@
 #include <string.h>
 
 // hack into header files that we can ship
-typedef int qboolean;
+typedef short qboolean;
 typedef unsigned char byte;
-#include "../utils/common/mathlib.h"
+#include "platform.h"
+#include "mathlib.h"
 #include "const.h"
 #include "progs.h"
 #include "progdefs.h"
